requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
monai>=1.3.0
timm>=0.9.12

# Data / DICOM
pydicom>=2.4.0
SimpleITK>=2.3.0
nibabel>=5.2.0
numpy>=1.26.0
scipy>=1.12.0
scikit-image>=0.22.0
pandas>=2.2.0

# Experiment management
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.16.0

# Web app
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
celery>=5.3.6
redis>=5.0.1
python-multipart>=0.0.9

# Visualisation
matplotlib>=3.8.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Export
onnx>=1.16.0
onnxruntime>=1.17.0
